Move IUserService and UserService to TimeScout.Application project
