+++ TimeScout.Application/Interfaces/IUserService.cs
@@ -1,13 +1,12 @@
-using System;
-using TimeScout.API.Models;
+using TimeScout.Domain.Entities;
 
-namespace TimeScout.API.Services;
+namespace TimeScout.Application.Interfaces;
 
 public interface IUserService
 {
     Task<bool> DeleteRefreshTokenAsync(int userId);
     Task<bool> DeleteUserAsync(int userId);
-    Task<User> GetUserByIdAsync(int userId);
+    Task<User?> GetUserByIdAsync(int userId);
     Task<bool> RecoverUserAsync(int userId);
-    Task<User> UpdateUserAsync(User user);
+    Task<User?> UpdateUserAsync(User user);
 }
